Commit Message: Refactor Jenkinsfile for new workflow and application requirements
Changed 'SERVICE_NAME', 'IMAGE_NAME', and 'REPOSITORY_NAME' values in the Jenkinsfile to reflect the new application usage of a different image and repository. Specifically, the original huawei-metrics-based image and repository was substituted with ones based on csctracker-ubuntu. Furthermore, eliminated 'Service update' section as it's not required in the new workflow format. This modification assists in streamlining and simplifying the build process, adapting it to the application's evolving needs.